kvspb_time_ticket/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-filter==25.1
django-grappelli==4.0.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.9
et_xmlfile==2.0.0
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
multidict==6.1.0
numpy==2.3.1
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.16.0
